Create modules folder to avoid circular dependencies
