+++ react/modules/getLabel.ts
@@ -1,0 +1,3 @@
+export default function getLabel(localeId: string) {
+  return localeId.split('-')[0]
+}
