+++ react/tsconfig.json
@@ -15,6 +15,6 @@
   "typeAcquisition": {
     "enable": false
   },
-  "include": ["./typings/*.d.ts", "./**/*.tsx"],
+  "include": ["./typings/*.d.ts", "./**/*.tsx", "./**/*.ts"],
   "exclude": ["node_modules"]
 }
