+++ react/typings/langs.d.ts
@@ -1,0 +1,6 @@
+declare module 'langs' {
+  export interface SupportedLanguage {
+    text: string
+    localeId: string
+  }
+}
